+++ NEWS.md
@@ -15,6 +15,8 @@
 
 * colwise functions `summarise_at()` ... can rename vars in the case of multiple functions (#4180).
 
+* `select_if()` and `rename_if()` handle logical vector predicate (#4213). 
+
 # dplyr 0.8.0.1 (2019-02-15)
 
 * Fixed integer C/C++ division, forced released by CRAN (#4185).